feat: create react button
