+++ src/components/button/index.js
@@ -1,0 +1,3 @@
+import { Button } from './styles';
+
+export default Button;
