+++ src/stories/button.stories.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import Button from '../components/button';
+
+export default {
+    title: 'Components/Button',
+    component: Button,
+};
+
+const Template = (args) => <Button {...args} />;
+
+export const Primary = Template.bind({});
+
+Primary.args = {
+    label: 'Lorem ipsum',
+    type: 'button',
+    variant: 'primary',
+};
